+++ router/router.js
@@ -1,0 +1,18 @@
+import Vue from "vue";
+import Router from "vue-router";
+
+Vue.use(Router);
+
+export default new Router({
+  mode: "history",
+  routes: [
+    {
+      path: "/post",
+      name: "post"
+    },
+    {
+      path: "/",
+      name: ""
+    }
+  ]
+});
